embedded file paths must always use forward-slash separator, even on windows. (#338)
Signed-off-by: Mike Cobbett <77053+techcobweb@users.noreply.github.com>

To fix this issue: Programming logic error is scary #2116

The warning was emitted because a folder path and file path were being joined together using the filepath.Join call. This draws the file separator from the operating system, rather than from the file system which it knows nothing about.

So on windows, this resulted in folder/file format. or folder\file on Windows.
But the embedded file system will only work with forward-slashes. folder/file
So ... now doing string concatenations with a separator derived from the read only file system, which is unaffected when running on Windows.